move storybook config into .storybook
