+++ .gitignore
@@ -1,10 +1,9 @@
 node_modules
-yarn-*.log
 docs/storybook-static
+docs/.storybook/preview.css
 .DS_Store
 dist
 .cache
 .yalc
 yalc.loc
 .turbo
-preview.css
